remove backstop from package.json and instal globally
