+++ backstop/lib/playwright/clickAndHoverHelper.cjs
