+++ backstop/lib/playwright/interceptImages.cjs
