+++ backstop/lib/playwright/loadCookies.cjs
