+++ backstop/lib/playwright/onBefore.cjs
@@ -1,4 +1,4 @@
 module.exports = async (page, scenario, viewport, isReference, browserContext) => {
   console.log('SCENARIO ON BEFORE > ' + scenario.label);
-  await require('./loadCookies')(browserContext, scenario, scenario.cookieConsent);
+  await require('./loadCookies.cjs')(browserContext, scenario, scenario.cookieConsent);
 };
