+++ backstop/lib/playwright/onReady.cjs
@@ -1,6 +1,6 @@
 module.exports = async (page, scenario, viewport, isReference, browserContext) => {
   console.log('SCENARIO ON READY > ', scenario.label, viewport.label);
-  await require('./clickAndHoverHelper')(page, scenario);
+  await require('./clickAndHoverHelper.cjs')(page, scenario);
 
   // add more ready handlers here...
 };
