+++ backstop/lib/setSessionCookie.cjs
